+++ Sources/VComponents/Components/Modals/Alert/VAlertButtonConvertible.swift
@@ -12,7 +12,6 @@ import SwiftUI
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
 @available(visionOS, unavailable)
-@MainActor
 public protocol VAlertButtonConvertible {
     /// Converts self to `VAlertButtonProtocol` `Array`.
     func toButtons() -> [any VAlertButtonProtocol]
@@ -29,5 +28,5 @@ extension Array: VAlertButtonConvertible where Element == any VAlertButtonProtoc
 @available(watchOS, unavailable)
 @available(visionOS, unavailable)
 extension EmptyView: VAlertButtonConvertible {
-    public func toButtons() -> [any VAlertButtonProtocol] { [] }
+    nonisolated public func toButtons() -> [any VAlertButtonProtocol] { [] }
 }